[hotfix] fix: 마일스톤 진행률 계산식 변경

+++ src/main/java/com/issuetracker/domain/milestone/response/MilestoneResponse.java
@@ -23,7 +23,7 @@ public class MilestoneResponse {
         Integer totalIssues = milestone.getTotalIssues();
         Integer openIssues = milestone.getOpenIssues();
         int closeIssues = totalIssues - openIssues;
-        int progress = closeIssues > 0 ? closeIssues / totalIssues * 100 : 0;
+        int progress = closeIssues > 0 ? (int) (((double) closeIssues / totalIssues) * 100) : 0;
 
         return MilestoneResponse.builder()
                 .id(milestone.getId())
